+++ cumulus/client/cirrus-executor/Cargo.toml
@@ -32,8 +32,6 @@ tracing = "0.1.23"
 thiserror = "1.0.29"
 tokio = "1.17.0"
 
-polkadot-overseer = { path = "../../../polkadot/node/overseer" }
-
 cirrus-block-builder = { path = "../block-builder" }
 cirrus-client-executor-gossip = { path = "../executor-gossip" }
 cirrus-primitives = { path = "../../primitives" }
