--- polkadot/README.md
@@ -1,3 +1,0 @@
-# Polkadot fork
-
-This is the fork of [polkadot](https://github.com/paritytech/polkadot/commit/c87a220490f88dfee0d5602fd834f4801916b269), required by the [culumus fork](../cumulus). The core part of this fork is the borrowed overseer and subsystems.
