--- polkadot/node/overseer/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "polkadot-overseer"
-version = "0.9.13"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.52"
-sc-client-api = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-sp-api = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-sp-blockchain = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-futures = "0.3.21"
-parking_lot = "0.12.0"
-thiserror = "1.0.30"
-tracing = "0.1.32"
-pin-project = "1.0.10"
-
-sp-consensus-slots = { version = "0.10.0-dev", git = "https://github.com/paritytech/substrate", rev = "c364008a6c7da8456e17967f55edf51e45146998" }
-sp-executor = { path = "../../../crates/sp-executor" }
-subspace-core-primitives = { version = "0.1.0", path = "../../../crates/subspace-core-primitives" }
-subspace-runtime-primitives = { path = "../../../crates/subspace-runtime-primitives" }
-
-[dev-dependencies]
-futures = { version = "0.3.21", features = ["thread-pool"] }
-
-[features]
-default = []
